backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
networkx==3.4.2
chromadb==0.5.23
google-genai==0.3.0
python-dotenv==1.0.1
pydantic==2.10.4
